+++ cvs/input/config_file/inference/xdit/mi3xx_pytorch_xdit_flux1_dev_single.json
