+++ cvs/input/config_file/inference/xdit/mi3xx_pytorch_xdit_flux2_dev_single.json
