+++ cvs/input/config_file/inference/xdit/mi3xx_pytorch_xdit_wan22_14b_diffusers_single.json
@@ -1,27 +1,19 @@
 {
     "config": {
         "container_image": "rocm/ufb-private:xdit-gfx942-py3.12-prereleases-rocm10.0.0rc3-2805c93",
-        "container_name": "wan22-benchmark-dist",
+        "container_name": "wan22-benchmark",
         "hf_token_file": "/home/{user-id}/.hf_token",
         "hf_home": "/home/{user-id}",
         "output_base_dir": "/home/{user-id}/cvs_outputs",
         "model_repo": "/data/models/Wan2.2-I2V-A14B-Diffusers",
         "model_rev": "",
-        "nnodes": 2,
-        "master_addr": "<changeme>",
-        "master_port": 29500,
-        "nccl_ib_hca": "<changeme>",
-        "nccl_socket_ifname": "<changeme>",
-        "gloo_socket_ifname": "<changeme>",
-        "nccl_ib_gid_index": "3",
-        "nccl_debug": "INFO",
         "container_config": {
             "device_list": [
                 "/dev/dri",
                 "/dev/kfd"
             ],
             "volume_dict": {
-                "/home/{user-id}/cvs/cvs/lib/inference/pytorch_xdit/scripts/wan_i2v_example.py": "/benchmark/wan_i2v_example.py"
+                "/home/{user-id}/cvs/cvs/lib/inference/xdit/scripts/wan_i2v_example.py": "/benchmark/wan_i2v_example.py"
             },
             "env_dict": {
                 "CVS_WAN_XFUSER_PYPACKAGES": "/shared/amdgpu/home/dl_dcgpu_aac_service_request_qle/pypackages"
@@ -48,7 +40,7 @@
             "compile": false,
             "torchrun_nproc": 8,
             "ulysses_size": 8,
-            "ring_size": 2,
+            "ring_size": 1,
             "expected_results": {
                 "auto": {
                     "max_avg_pipe_time_s": 300.0
