+++ cvs/input/config_file/inference/xdit/mi3xx_pytorch_xdit_wan22_14b_single.json
