+++ cvs/lib/inference/xdit/scripts/wan_i2v_example.py
