+++ cvs/tests/inference/xdit/pytorch_xdit_wan22_14b_diffusers_distributed.py
@@ -24,15 +24,15 @@ from cvs.lib.utils_lib import (
 from cvs.lib import docker_lib
 from cvs.lib import globals
 from cvs.parsers.schemas import ClusterConfigFile, PytorchXditWanConfigFile
-from cvs.lib.inference.pytorch_xdit.pytorch_xdit_model_verify import (
+from cvs.lib.inference.xdit.pytorch_xdit_model_verify import (
     build_diffusers_local_model_required_checks,
     verify_required_checks_on_nodes,
 )
-from cvs.lib.inference.pytorch_xdit.pytorch_xdit_wan_i2v import (
+from cvs.lib.inference.xdit.pytorch_xdit_wan_i2v import (
     WanI2vOutputParser,
     log_results_summary,
 )
-from cvs.lib.inference.pytorch_xdit.pytorch_xdit_wan_job import (
+from cvs.lib.inference.xdit.pytorch_xdit_wan_job import (
     build_wan_output_cleanup_cmd,
     compute_world_size,
     launch_wan_benchmark,
