+++ cvs/tests/inference/xdit/pytorch_xdit_wan22_14b_single.py
@@ -30,14 +30,14 @@ from cvs.lib.utils_lib import (
 from cvs.lib import docker_lib
 from cvs.lib import globals
 from cvs.parsers.schemas import ClusterConfigFile, PytorchXditWanConfigFile
-from cvs.lib.inference.pytorch_xdit.pytorch_xdit_model_verify import (
+from cvs.lib.inference.xdit.pytorch_xdit_model_verify import (
     build_diffusers_local_model_required_checks,
     resolve_wan_local_model_required_checks,
     verify_required_checks_on_nodes,
     wan_native_hf_snapshot_required_checks,
 )
-from cvs.lib.inference.pytorch_xdit.pytorch_xdit_wan import WanOutputParser
-from cvs.lib.inference.pytorch_xdit.pytorch_xdit_wan_job import (
+from cvs.lib.inference.xdit.pytorch_xdit_wan import WanOutputParser
+from cvs.lib.inference.xdit.pytorch_xdit_wan_job import (
     launch_wan_benchmark,
     store_resolved_wan_model_format_from_index,
 )
